# backend/openapi.json
{
  "openapi": "3.2.0",
  "info": {
    "title": "WattPrint API",
    "summary": "Household energy disaggregation (NILM) for WattPrint",
    "description": "Whole-house power in, appliance-level consumption out.\n\n* **Demo** — one household (Plegma House 101), every result precomputed: no model runs, every\n  call answers in milliseconds. Start with `GET /api/v1/demo/household`.\n* **Platform** — devices, raw readings, CSV import of NILM results and live disaggregation.\n\nErrors follow RFC 9457 (`application/problem+json`). Times are ISO 8601, UTC.\n",
    "version": "0.2.0"
  },
  "servers": [
    {
      "url": "/",
      "description": "The host serving this document",
      "name": "current"
    }
  ],
  "paths": {
...
    "/api/v1/devices": {
      "get": {
        "tags": [
          "devices"
        ],
        "summary": "List Devices",
        "operationId": "listDevices",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/DeviceRead"
                  },
                  "type": "array",
                  "title": "Response Listdevices"
                }
              }
            }
          },
          "default": {
            "summary": "Error",
            "description": "Any other error, as RFC 9457 problem details",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "devices"
        ],
        "summary": "Create Device",
        "operationId": "createDevice",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DeviceCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeviceRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationProblem"
                }
              }
            },
            "summary": "Invalid parameters"
          },
          "default": {
            "summary": "Error",
            "description": "Any other error, as RFC 9457 problem details",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/devices/{device_id}": {
      "get": {
        "tags": [
          "devices"
        ],
        "summary": "Get Device",
        "operationId": "getDevice",
        "parameters": [
          {
            "name": "device_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Device Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeviceRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationProblem"
                }
              }
            },
            "summary": "Invalid parameters"
          },
          "default": {
            "summary": "Error",
            "description": "Any other error, as RFC 9457 problem details",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/devices/{device_id}/readings": {
      "post": {
        "tags": [
          "readings"
        ],
        "summary": "Ingest",
        "operationId": "createReadings",
        "parameters": [
          {
            "name": "device_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Device Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReadingsBulkIn"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationProblem"
                }
              }
            },
            "summary": "Invalid parameters"
          },
          "default": {
            "summary": "Error",
            "description": "Any other error, as RFC 9457 problem details",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "readings"
        ],
        "summary": "Query",
        "operationId": "listReadings",
        "parameters": [
          {
            "name": "device_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Device Id"
            }
          },
          {
            "name": "start",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date-time",
              "title": "Start"
            }
          },
          {
            "name": "end",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date-time",
              "title": "End"
            }
          },
          {
            "name": "bucket",
            "in": "query",
            "required": false,
            "schema": {
              "enum": [
                "1 minute",
                "10 minutes",
                "1 hour",
                "1 day"
              ],
              "type": "string",
              "default": "1 hour",
              "title": "Bucket"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/BucketOut"
                  },
                  "title": "Response Listreadings"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationProblem"
                }
              }
            },
            "summary": "Invalid parameters"
          },
          "default": {
            "summary": "Error",
            "description": "Any other error, as RFC 9457 problem details",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/households": {
      "get": {
        "tags": [
          "disaggregation"
        ],
        "summary": "List Households",
        "operationId": "listHouseholds",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/HouseholdRead"
                  },
                  "type": "array",
                  "title": "Response Listhouseholds"
                }
              }
            }
          },
          "default": {
            "summary": "Error",
            "description": "Any other error, as RFC 9457 problem details",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/households/import": {
      "post": {
        "tags": [
          "disaggregation"
        ],
        "summary": "Import Disaggregation",
        "operationId": "importHouseholdDisaggregation",
        "parameters": [
          {
            "name": "household",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Household"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_importHouseholdDisaggregation"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ImportResult"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationProblem"
                }
              }
            },
            "summary": "Invalid parameters"
          },
          "default": {
            "summary": "Error",
            "description": "Any other error, as RFC 9457 problem details",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/households/{household_id}/disaggregation": {
      "get": {
        "tags": [
          "disaggregation"
        ],
        "summary": "Get Disaggregation",
        "operationId": "getHouseholdDisaggregation",
        "parameters": [
          {
            "name": "household_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Household Id"
            }
          },
          {
            "name": "start",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date-time",
              "title": "Start"
            }
          },
          {
            "name": "end",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date-time",
              "title": "End"
            }
          },
          {
            "name": "bucket",
            "in": "query",
            "required": false,
            "schema": {
              "enum": [
                "1 minute",
                "10 minutes",
                "1 hour",
                "1 day"
              ],
              "type": "string",
              "default": "1 hour",
              "title": "Bucket"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DisaggregationOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationProblem"
                }
              }
            },
            "summary": "Invalid parameters"
          },
          "default": {
            "summary": "Error",
            "description": "Any other error, as RFC 9457 problem details",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/disaggregate": {
      "post": {
        "tags": [
          "disaggregation"
        ],
        "summary": "Disaggregate",
        "description": "Whole-house power in, per-appliance power out. Proxies to the NILM inference service.",
        "operationId": "disaggregateLive",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DisaggregateIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationProblem"
                }
              }
            },
            "summary": "Invalid parameters"
          },
          "default": {
            "summary": "Error",
            "description": "Any other error, as RFC 9457 problem details",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Body_importHouseholdDisaggregation": {
        "properties": {
          "file": {
            "type": "string",
            "contentMediaType": "application/octet-stream",
            "title": "File",
            "description": "CSV from nilmformer-experiment outputs"
          }
        },
        "type": "object",
        "required": [
          "file"
        ],
        "title": "Body_importHouseholdDisaggregation"
      },
...
      "BucketOut": {
        "properties": {
          "bucket": {
            "type": "string",
            "format": "date-time",
            "title": "Bucket"
          },
          "avg_power_w": {
            "type": "number",
            "title": "Avg Power W"
          },
          "max_power_w": {
            "type": "number",
            "title": "Max Power W"
          },
          "energy_wh": {
            "type": "number",
            "title": "Energy Wh"
          }
        },
        "type": "object",
        "required": [
          "bucket",
          "avg_power_w",
          "max_power_w",
          "energy_wh"
        ],
        "title": "BucketOut"
      },
...
      "DeviceCreate": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "kind": {
            "$ref": "#/components/schemas/DeviceKind"
          },
          "parent_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Parent Id"
          }
        },
        "type": "object",
        "required": [
          "name",
          "kind"
        ],
        "title": "DeviceCreate"
      },
      "DeviceKind": {
        "type": "string",
        "enum": [
          "aggregate",
          "appliance"
        ],
        "title": "DeviceKind"
      },
      "DeviceRead": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "kind": {
            "$ref": "#/components/schemas/DeviceKind"
          },
          "parent_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Parent Id"
          },
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "name",
          "kind",
          "id",
          "created_at"
        ],
        "title": "DeviceRead"
      },
      "DisaggregateIn": {
        "properties": {
          "start": {
            "type": "string",
            "format": "date-time",
            "title": "Start"
          },
          "power_w": {
            "items": {
              "anyOf": [
                {
                  "type": "number"
                },
                {
                  "type": "null"
                }
              ]
            },
            "type": "array",
            "title": "Power W"
          },
          "stride": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Stride"
          }
        },
        "type": "object",
        "required": [
          "start",
          "power_w"
        ],
        "title": "DisaggregateIn"
      },
      "DisaggregationOut": {
        "properties": {
          "household_id": {
            "type": "string",
            "format": "uuid",
            "title": "Household Id"
          },
          "start": {
            "type": "string",
            "format": "date-time",
            "title": "Start"
          },
          "end": {
            "type": "string",
            "format": "date-time",
            "title": "End"
          },
          "bucket": {
            "type": "string",
            "title": "Bucket"
          },
          "totals": {
            "items": {
              "$ref": "#/components/schemas/TotalOut"
            },
            "type": "array",
            "title": "Totals"
          },
          "series": {
            "items": {
              "$ref": "#/components/schemas/SeriesOut"
            },
            "type": "array",
            "title": "Series"
          }
        },
        "type": "object",
        "required": [
          "household_id",
          "start",
          "end",
          "bucket",
          "totals",
          "series"
        ],
        "title": "DisaggregationOut"
      },
...
      "HouseholdRead": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "appliances": {
            "items": {
              "$ref": "#/components/schemas/DeviceRead"
            },
            "type": "array",
            "title": "Appliances"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "appliances"
        ],
        "title": "HouseholdRead"
      },
      "ImportResult": {
        "properties": {
          "household_id": {
            "type": "string",
            "format": "uuid",
            "title": "Household Id"
          },
          "rows_per_device": {
            "additionalProperties": {
              "type": "integer"
            },
            "type": "object",
            "title": "Rows Per Device"
          }
        },
        "type": "object",
        "required": [
          "household_id",
          "rows_per_device"
        ],
        "title": "ImportResult"
      },
...
      "Problem": {
        "additionalProperties": true,
        "description": "RFC 9457 problem details. Extension members may be added.",
        "properties": {
          "type": {
            "default": "about:blank",
            "description": "URI reference identifying the problem type",
            "title": "Type",
            "type": "string"
          },
          "title": {
            "description": "Short, human-readable summary of the problem type",
            "examples": [
              "Not Found"
            ],
            "title": "Title",
            "type": "string"
          },
          "status": {
            "description": "HTTP status code",
            "examples": [
              404
            ],
            "title": "Status",
            "type": "integer"
          },
          "detail": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "description": "Explanation specific to this occurrence",
            "title": "Detail"
          },
          "instance": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "description": "URI reference of this occurrence (the path)",
            "title": "Instance"
          }
        },
        "required": [
          "title",
          "status"
        ],
        "title": "Problem",
        "type": "object"
      },
...
      "ReadingIn": {
        "properties": {
          "time": {
            "type": "string",
            "format": "date-time",
            "title": "Time"
          },
          "power_w": {
            "type": "number",
            "title": "Power W"
          }
        },
        "type": "object",
        "required": [
          "time",
          "power_w"
        ],
        "title": "ReadingIn"
      },
      "ReadingsBulkIn": {
        "properties": {
          "readings": {
            "items": {
              "$ref": "#/components/schemas/ReadingIn"
            },
            "type": "array",
            "title": "Readings"
          }
        },
        "type": "object",
        "required": [
          "readings"
        ],
        "title": "ReadingsBulkIn"
      },
...
      "SeriesOut": {
        "properties": {
          "device_id": {
            "type": "string",
            "format": "uuid",
            "title": "Device Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "kind": {
            "type": "string",
            "title": "Kind"
          },
          "points": {
            "items": {
              "$ref": "#/components/schemas/BucketOut"
            },
            "type": "array",
            "title": "Points"
          }
        },
        "type": "object",
        "required": [
          "device_id",
          "name",
          "kind",
          "points"
        ],
        "title": "SeriesOut"
      },
      "TotalOut": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "energy_wh": {
            "type": "number",
            "title": "Energy Wh"
          },
          "share_pct": {
            "type": "number",
            "title": "Share Pct"
          }
        },
        "type": "object",
        "required": [
          "name",
          "energy_wh",
          "share_pct"
        ],
        "title": "TotalOut"
      },
      "ValidationProblem": {
        "additionalProperties": true,
        "properties": {
          "type": {
            "default": "about:blank",
            "description": "URI reference identifying the problem type",
            "title": "Type",
            "type": "string"
          },
          "title": {
            "description": "Short, human-readable summary of the problem type",
            "examples": [
              "Not Found"
            ],
            "title": "Title",
            "type": "string"
          },
          "status": {
            "description": "HTTP status code",
            "examples": [
              404
            ],
            "title": "Status",
            "type": "integer"
          },
          "detail": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "description": "Explanation specific to this occurrence",
            "title": "Detail"
          },
          "instance": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "description": "URI reference of this occurrence (the path)",
            "title": "Instance"
          },
          "errors": {
            "description": "One entry per invalid input: `loc` (where), `msg` (why), `type` (error code)",
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "title": "Errors",
            "type": "array"
          }
        },
        "required": [
          "title",
          "status",
          "errors"
        ],
        "title": "ValidationProblem",
        "type": "object"
      },
...
    }
  },
...
}